refactor: extract useScrollRestoration primitive from SearchList
